requirements.txt
Django==4.1.2
psycopg2-binary==2.9.4
djangorestframework==3.14.0
